Refactored integration tests (#34)
