+++ src/test/java/dev/fastgql/integration/ContainerEnvWithPostgres.java
@@ -1,0 +1,36 @@
+package dev.fastgql.integration;
+
+import dev.fastgql.db.DatasourceConfig;
+import io.debezium.testing.testcontainers.ConnectorConfiguration;
+import org.testcontainers.containers.JdbcDatabaseContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+public class ContainerEnvWithPostgres extends AbstractContainerEnvImpl {
+  @Override
+  protected JdbcDatabaseContainer<?> createJdbcContainer() {
+    return new PostgreSQLContainer<>("debezium/postgres:11")
+        .withNetwork(network)
+        .withNetworkAliases("postgres");
+  }
+
+  @Override
+  protected ConnectorConfiguration createConnectorConfiguration() {
+    return ConnectorConfiguration.forJdbcContainer(jdbcDatabaseContainer)
+        .with("database.server.name", "dbserver")
+        .with("slot.name", "debezium");
+  }
+
+  @Override
+  protected DatasourceConfig createDatasourceConfig() {
+    return DatasourceConfig.createDatasourceConfig(
+        jdbcDatabaseContainer.getJdbcUrl(),
+        jdbcDatabaseContainer.getUsername(),
+        jdbcDatabaseContainer.getPassword(),
+        "public");
+  }
+
+  @Override
+  protected String getJdbcUrlForMultipleQueries() {
+    return jdbcDatabaseContainer.getJdbcUrl();
+  }
+}
