+++ src/test/java/dev/fastgql/integration/SetupTearDownForEach.java
@@ -1,0 +1,17 @@
+package dev.fastgql.integration;
+
+import io.vertx.junit5.VertxTestContext;
+import io.vertx.reactivex.core.Vertx;
+import org.junit.jupiter.api.*;
+
+public interface SetupTearDownForEach extends AbstractContainerEnv {
+  @BeforeEach
+  default void beforeEach(Vertx vertx, VertxTestContext context) {
+    setup(vertx, context);
+  }
+
+  @AfterEach
+  default void afterEach(Vertx vertx, VertxTestContext context) {
+    tearDown(vertx, context);
+  }
+}
